Reformat documentation comments
